% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{add_router_callback}
\alias{add_router_callback}
\title{Set a router callback function}
\usage{
add_router_callback(router, callback = c("pageload", "default"), func)
}
\arguments{
\item{router}{A router object.}

\item{callback}{The name of the callback to set. Currently \code{"pageload"} and
\code{"default"} callbacks are supported.}

\item{func}{The callback function. The function should take a \code{router_page}
object and shiny session as input.}
}
\value{
Invisibly returns the input \code{router} object.
}
\description{
Set a router callback function
}
\examples{
\dontrun{
add_router_callback(
  router,
  "pageload",
  function(page, session) {
    shinyjs::runjs(paste0("document.title = \"", page$title, "\";"))
  }
)

add_router_callback(
  router,
  "default",
  function(session) {
    if ("dev" in session$groups) {
      "devlanding"
    } else {
      "home"
    }
  }
)
}
}
