% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{change_router_page}
\alias{change_router_page}
\title{Change router page}
\usage{
change_router_page(
  path,
  mode = "push",
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{path}{Path of the destination page.}

\item{mode}{Argument passed to \code{\link[shiny:updateQueryString]{shiny::updateQueryString()}}.}

\item{session}{The current shiny session object.}
}
\description{
Change router page
}
\examples{
\dontrun{
change_router_page("home")
}
}
