% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{get_router_page_metadata}
\alias{get_router_page_metadata}
\title{Get the metadata of the current page}
\usage{
get_router_page_metadata(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{The current shiny session object.}
}
\value{
The title of the current \code{\link[=router_page]{router_page()}}.
}
\description{
Get the list stored in the metadata attribute of the current page.
}
\examples{
\dontrun{
get_router_page_title()
}
}
