% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{get_router_page_title}
\alias{get_router_page_title}
\title{Get the title of the current router page}
\usage{
get_router_page_title(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{The current shiny session object.}
}
\value{
The title of the current \code{\link[=router_page]{router_page()}}.
}
\description{
Get the title of the current router page
}
\examples{
\dontrun{
get_router_page_title()
}
}
