% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{get_router_path}
\alias{get_router_path}
\title{Get the current router path from the URL}
\usage{
get_router_path(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{The current shiny session object.}
}
\value{
The current router path.
}
\description{
Get the current router path from the URL
}
\examples{
\dontrun{
get_router_path()
}
}
