% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{router_link}
\alias{router_link}
\title{Created a page link}
\usage{
router_link(path)
}
\arguments{
\item{path}{Path of the destination page.}
}
\value{
A link to the destination page.
}
\description{
Take a path to a page and convert it to a link for use in Shiny UI.
}
\examples{
htmltools::a(href = router_link("home"), "Home page")
}
