% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{router_server}
\alias{router_server}
\title{Router server function}
\usage{
router_server(router, input, output, session, id = "__router__")
}
\arguments{
\item{router}{A router object.}

\item{input}{Shiny server \code{input} object.}

\item{output}{Shiny server \code{output} object.}

\item{session}{Shiny server \code{session} object.}

\item{id}{Router id.}
}
\value{
A shiny observer reference class object. See \code{\link[shiny:observe]{shiny::observe()}} for
details.
}
\description{
Register server code. This should be placed in your shiny server, and have a
corresponding \code{\link[=router_ui]{router_ui()}} element.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {

ui_test <- function(id) {
  htmltools::div(
    htmltools::p("Test!"),
    htmltools::p(htmltools::a(href = router_link("/"), "Home"))
  )
}

test_router <- router(router_page("home", ui_test))

ui <- shiny::fluidPage(
  router_ui()
)

server <- function(input, output, session) {
  router_server(test_router, input, output, session)
}

shinyApp(ui, server)
}
}
