% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{router_ui}
\alias{router_ui}
\title{Router UI output}
\usage{
router_ui(id = "__router__")
}
\arguments{
\item{id}{Router id.}
}
\value{
An HTML output element.
}
\description{
Placeholder that will be used to render the router page. This can be placed
anywhere within a shiny UI.
}
