% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{get_router_page}
\alias{get_router_page}
\title{Get the current \code{router_page} object}
\usage{
get_router_page(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{The current shiny session object.}
}
\value{
The current \code{\link[=router_page]{router_page()}} object.
}
\description{
Get the current \code{router_page} object
}
\examples{
\dontrun{
get_router_page()
}
}
