% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{router}
\alias{router}
\title{Create a router}
\usage{
router(
  ...,
  callback_pageload = NULL,
  callback_default = NULL,
  page_403 = router_page("403", ui_403),
  page_404 = router_page("404", ui_404)
)
}
\arguments{
\item{...}{A set of \code{\link[=router_page]{router_page()}} objects. The first page is considered the
default, and will be used if no path is provided.}

\item{callback_pageload}{A function that is called immediately after each
page is rendered. The function should take a \code{router_page} object and shiny
session as input.}

\item{callback_default}{A function that determines the default page. The
function should take a shiny session as input and return the path of the
default page. If not provided, the first page is used as the default.}

\item{page_403}{A \code{router_page} to be used when a user attempts to access an
unauthorised page.}

\item{page_404}{A \code{router_page} to be used when a user attempts to access a
page that doesn't exist.}
}
\value{
A \code{router} object.
}
\description{
Create a router
}
\examples{
if (interactive()) {

ui_test <- function(id) {
  htmltools::p("Test!")
}

home_page <- router_page("home", ui_test)
other_page - router_page("other", ui_test)

test_router <- router(home_page, other_page)
}
}
